move requestFrame from common to drawing.
